Updated: es5 to es6 (modules & classes)
