+++ gulpfile.js
@@ -13,8 +13,18 @@ gulp.task(
     pkg.name + '/build',
     function() {
         return gulp
-            .src('./src/module.js')
+            .src('./src/index.js')
             .pipe(webpackStream({
+                module: {
+                    loaders: [
+                        // https://github.com/babel/babel-loader
+                        {test: /\.js$/, loader: 'babel'},
+                        // https://github.com/webpack/json-loader
+                        {test: /\.json$/, loader: 'json'},
+                        // https://github.com/webpack/html-loader
+                        {test: /\.html$/, loader: 'html'}
+                    ]
+                },
                 plugins: [
                     // http://webpack.github.io/docs/list-of-plugins.html#uglifyjsplugin
                     new webpack.optimize.UglifyJsPlugin({
