+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "angular-file-upload",
-    "version": "1.1.8",
+    "version": "1.2.0",
     "homepage": "https://github.com/nervgh/angular-file-upload",
     "description": "Angular File Upload is a module for the AngularJS framework",
     "author": {
@@ -9,6 +9,10 @@
     },
     "main": "dist/angular-file-upload.min.js",
     "devDependencies": {
+        "babel": "^4.7.16",
+        "babel-loader": "^4.0.0",
+        "json-loader": "^0.5.1",
+        "html-loader": "^0.2.3",
         "gulp": "^3.9.0",
         "webpack": "^1.10.1",
         "webpack-stream": "^2.0.0"
